third_party/blink/web_tests/fast/replaced/table-percent-height-text-controls-expected.txt

This test checks that text controls with percentage heights within table cells have the correct height.Text controls are in a different test than other replaced elements because their metrics are platform-specific.The reason a 75% control is the same height as a 100% control is because a replaced element that depends on theheight of its parent cell is treated as auto. So by itself it will set the height of the row. See https://drafts.csswg.org/css-tables-3/#row-layout

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS getWidth('input-password-75') is getWidth('input-password-100')
PASS getFullHeight('input-password-75') != '0px' is true
PASS getFullHeight('input-password-75') is getFullHeight('input-password-100')
PASS getWidth('input-text-75') is getWidth('input-text-100')
PASS getFullHeight('input-text-75') != '0px' is true
PASS getFullHeight('input-text-75') is getFullHeight('input-text-100')
PASS getWidth('textarea-75') is getWidth('textarea-100')
PASS getFullHeight('textarea-75') != '0px' is true
PASS getFullHeight('textarea-75') is getFullHeight('textarea-100')
PASS successfullyParsed is true

TEST COMPLETE
